{"schema":"decodex/commit/1","summary":"Extract input cursor movement owner","authority":"manual"}
